moved classes to proper module
