+++ lib/rubies/random_array.rb
@@ -1,0 +1,25 @@
+module Rubies
+  class RandomArray < ::Array
+
+    def mini_array
+      (-1_000..1_000).sort_by { rand }.sample 3
+    end
+
+    def nesting_array
+      result = []
+      rand(1..3).times do
+        result << mini_array
+      end
+      result.each do |array|
+        array << self.mini_array
+      end
+      result
+    end
+
+    def generate
+      depth = rand(0..3)
+      nesting_array.flatten(depth)
+    end
+
+  end
+end
